# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VetHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vet/search-owner:
    get:
      operationId: searchPetOwner
      tags: [pets]
      summary: Search pet owner by phone number
      parameters:
        - name: phone
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Owner and their pets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerWithPets"
  /vet/search-pet:
    get:
      operationId: searchPet
      tags: [pets]
      summary: Search pets by name (vet)
      parameters:
        - name: name
          in: query
          required: true
          schema: { type: string }
        - name: clinicId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Matching pets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PetWithOwner"
  /vet/add-pet-for-owner:
    post:
      operationId: addPetForOwner
      tags: [pets]
      summary: Vet adds a pet linked to an owner by phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VetAddPetInput"
      responses:
        "201":
          description: Created pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"

  # ── Vaccinations ──────────────────────────────────────────────────────────
...
  /visits/{visitId}:
    get:
      operationId: getVisit
      tags: [visits]
      summary: Get a visit by ID with all details
      parameters:
        - name: visitId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Visit with full details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitDetail"
    patch:
      operationId: updateVisit
      tags: [visits]
      summary: Update visit (anamnesis, therapy, status, etc.)
      parameters:
        - name: visitId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitUpdate"
      responses:
        "200":
          description: Updated visit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitDetail"
  /visits/{visitId}/items:
    post:
      operationId: addVisitItem
      tags: [visits]
      summary: Add a service, medicine, supporting item, or custom item to a visit
      parameters:
        - name: visitId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitItemInput"
      responses:
        "201":
          description: Created visit item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitItem"
  /visit-items/{itemId}:
    patch:
      operationId: updateVisitItem
      tags: [visits]
      summary: Update a visit item
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitItemUpdate"
      responses:
        "200":
          description: Updated visit item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitItem"
    delete:
      operationId: deleteVisitItem
      tags: [visits]
      summary: Remove a visit item
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Daily Reports (inpatient) ─────────────────────────────────────────────
  /visits/{visitId}/daily-reports:
    get:
      operationId: listDailyReports
      tags: [daily-reports]
      summary: List daily reports for an inpatient visit
      parameters:
        - name: visitId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Daily reports list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyReport"
    post:
      operationId: createDailyReport
      tags: [daily-reports]
      summary: Add a daily inpatient report
      parameters:
        - name: visitId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyReportInput"
      responses:
        "201":
          description: Created daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id: { type: integer }
        stackId: { type: string }
        name: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        role:
          type: string
          enum: [pet_owner, vet, vet_owner, both]
        isPetOwner: { type: boolean }
        isVet: { type: boolean }
        isVetOwner: { type: boolean }
        clinicId: { type: ["integer", "null"] }
        createdAt: { type: string }
      required: [id, stackId, role, isPetOwner, isVet, isVetOwner, createdAt]
...
    Pet:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        dateOfBirth: { type: ["string", "null"] }
        gender:
          type: string
          enum: [male, female, unknown]
        sterilized: { type: boolean }
        color: { type: ["string", "null"] }
        speciesId: { type: integer }
        speciesName: { type: ["string", "null"] }
        ownerId: { type: integer }
        ownerName: { type: ["string", "null"] }
        ownerPhone: { type: ["string", "null"] }
        status:
          type: string
          enum: [healthy, sick, hospitalized, need_intensive_care, passed_away]
        photoUrl: { type: ["string", "null"] }
        createdAt: { type: string }
      required:
        [id, name, gender, sterilized, speciesId, ownerId, status, createdAt]
...
    PetWithOwner:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        speciesName: { type: ["string", "null"] }
        gender: { type: string }
        status: { type: string }
        color: { type: ["string", "null"] }
        photoUrl: { type: ["string", "null"] }
        ownerId: { type: integer }
        ownerName: { type: ["string", "null"] }
        ownerPhone: { type: ["string", "null"] }
      required: [id, name, gender, status, ownerId]
    OwnerWithPets:
      type: object
      properties:
        owner:
          $ref: "#/components/schemas/UserProfile"
        pets:
          type: array
          items:
            $ref: "#/components/schemas/Pet"
      required: [owner, pets]
    VetAddPetInput:
      type: object
      properties:
        ownerPhone: { type: string }
        name: { type: string, minLength: 1 }
        dateOfBirth: { type: string }
        gender:
          type: string
          enum: [male, female, unknown]
        sterilized: { type: boolean }
        color: { type: string }
        speciesId: { type: integer }
      required: [ownerPhone, name, gender, sterilized, speciesId]
...
    VisitDetail:
      type: object
      properties:
        id: { type: integer }
        petId: { type: integer }
        petName: { type: ["string", "null"] }
        clinicId: { type: integer }
        vetId: { type: ["integer", "null"] }
        vetName: { type: ["string", "null"] }
        type: { type: string }
        status: { type: string }
        anamnesis: { type: ["string", "null"] }
        therapy: { type: ["string", "null"] }
        visitDate: { type: string }
        dischargeDate: { type: ["string", "null"] }
        deposit: { type: ["number", "null"] }
        totalCost: { type: number }
        billedCost: { type: number }
        items:
          type: array
          items:
            $ref: "#/components/schemas/VisitItem"
        dailyReports:
          type: array
          items:
            $ref: "#/components/schemas/DailyReport"
        createdAt: { type: string }
      required:
        [
          id,
          petId,
          clinicId,
          type,
          status,
          visitDate,
          totalCost,
          billedCost,
          items,
          dailyReports,
          createdAt,
        ]
...
    VisitUpdate:
      type: object
      properties:
        anamnesis: { type: string }
        therapy: { type: string }
        status:
          type: string
          enum: [active, completed, cancelled]
        dischargeDate: { type: string }
        vetId: { type: integer }
        deposit: { type: number }
    VisitItem:
      type: object
      properties:
        id: { type: integer }
        visitId: { type: integer }
        category:
          type: string
          enum: [service, medicine, supporting, other]
        name: { type: string }
        description: { type: ["string", "null"] }
        quantity: { type: number }
        unitPrice: { type: number }
        totalPrice: { type: number }
        itemDate: { type: string }
        isPaid: { type: boolean }
      required:
        [
          id,
          visitId,
          category,
          name,
          quantity,
          unitPrice,
          totalPrice,
          itemDate,
          isPaid,
        ]
    VisitItemInput:
      type: object
      properties:
        category:
          type: string
          enum: [service, medicine, supporting, other]
        name: { type: string, minLength: 1 }
        description: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
        itemDate: { type: string, description: "ISO date YYYY-MM-DD" }
        isPaid: { type: boolean }
      required: [category, name, quantity, unitPrice, itemDate]
    VisitItemUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
        isPaid: { type: boolean }
    DailyReport:
      type: object
      properties:
        id: { type: integer }
        visitId: { type: integer }
        reportDate: { type: string }
        condition: { type: ["string", "null"] }
        treatment: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        cost: { type: number }
        vetName: { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, visitId, reportDate, cost, createdAt]
    DailyReportInput:
      type: object
      properties:
        reportDate: { type: string }
        condition: { type: string }
        treatment: { type: string }
        notes: { type: string }
        cost: { type: number }
      required: [reportDate, cost]
...